LLL attack with PDF documentation.
Academic task to do. We had to simulate breaking the NTRU ciphertext and
give conclusion about results.

It's written strictly for GNU C++ compiler because the same project
(with few  improvments like adding vector of vector in 6 six lines
instead of 36, thanks to MSVC C++ compiler) in MSVC's been running for 7
minutes, and in GNU it's been about few seconds. Makes a difference,
doesn't it?

I've got a B+/A- for this (4,5).